Changes to add vms
fixed some errors,
you couldn't get a join token if the first vm was already in the swarm.
mostly decomposition again, some  other changes that were probably unnecessary really like checking if the repository/docker are installed. Unless we want to have it be able to update the repository so we don't have to recreate instances constantly